feat: add reset board
