+++ src/components/App/App.js
@@ -1,25 +1,13 @@
 import React from 'react';
-// import React, { useEffect, useState } from 'react';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import './App.css';
 import Header from '../Header/Header';
 import store from '../../redux/store';
-// import Board from '../Board/Board';
-// import initializeDeck from '../../deck';
 import Main from '../Main/Main';
-import MainContainer from '../Main/MainContainer';
+// import BoardContainer from '../Board/BoardContainer';
 
 function App() {
-  // const [cards, setCards] = useState([]);
-
-  // const [flipped, setFlipped] = useState([]);
-
-  // const handleClick = (id) => setFlipped([...flipped, id]);
-
-  // useEffect(() => {
-  //   setCards(initializeDeck())
-  // },[])
 
   return (
     <Provider store={store}>
