+++ src/components/Board/boardSelector.js
@@ -1,0 +1,5 @@
+export const getCards = (state) => state.main.cards;
+export const getFlipped = (state) => state.main.flipped;
+export const getDimension = (state) => state.main.dimension;
+export const getDisabled = (state) => state.main.disabled;
+export const getSolved = (state) => state.main.solved;
