+++ src/components/Main/Main.module.css
@@ -1,0 +1,9 @@
+.main{
+   display: grid;
+   grid-gap: 20px;
+   justify-items: center;
+}
+
+.btn{
+   width: 30%;
+}
