+++ src/redux/mainReducer.js
@@ -5,6 +5,7 @@ const MAIN_SET_FLIPPED = 'MAIN_SET_FLIPPED';
 const MAIN_SET_DIMENSION = 'MAIN_SET_DIMENSION';
 const MAIN_SET_SOLVED = 'MAIN_SET_SOLVED';
 const MAIN_SET_DISABLED = 'MAIN_SET_DISABLED';
+const MAIN_RESET_STATE = 'MAIN_RESET_STATE';
 
 const initialState = {
   cards: initializeDeck(),
@@ -12,18 +13,28 @@ const initialState = {
   dimension: 480,
   solved: [],
   disabled: false,
+  countPairs: 5,
 };
 
 const mainReducer = (state = initialState, action) => {
-  switch (action) {
+  switch (action.type) {
+    case MAIN_RESET_STATE:
+      return {
+        ...state,
+        cards: initializeDeck(action.countPairs),
+        flipped: [],
+        dimension: 480,
+        solved: [],
+        disabled: false,
+      };
+
     case MAIN_SET_CARDS:
       return {
         ...state,
         cards: [...action.cards],
       };
 
-     case MAIN_SET_FLIPPED:
-        debugger
+    case MAIN_SET_FLIPPED:
       return {
         ...state,
         flipped: [...action.array],
@@ -32,7 +43,7 @@ const mainReducer = (state = initialState, action) => {
     case MAIN_SET_DIMENSION:
       return {
         ...state,
-        dimension: [...action.value],
+        dimension: action.value,
       };
 
     case MAIN_SET_SOLVED:
@@ -41,8 +52,7 @@ const mainReducer = (state = initialState, action) => {
         solved: [...action.array],
       };
 
-     case MAIN_SET_DISABLED:
-        debugger
+    case MAIN_SET_DISABLED:
       return {
         ...state,
         disabled: action.boolean,
@@ -53,17 +63,15 @@ const mainReducer = (state = initialState, action) => {
   }
 };
 
-export const setCards = (cards) => ({ action: MAIN_SET_CARDS, cards });
-export const setFlipped = (array) => {
-   debugger
-   return { action: MAIN_SET_FLIPPED, array }
-};
-export const setDimension = (value) => ({ action: MAIN_SET_DIMENSION, value });
-export const setSolved = (array) => ({ action: MAIN_SET_SOLVED, array });
-export const setDisabled = (boolean) => {
-   debugger
-   return { action: MAIN_SET_DISABLED, boolean }
-};
+export const resetState = (countPairs) => ({
+  type: MAIN_RESET_STATE,
+  countPairs,
+});
+export const setCards = (cards) => ({ type: MAIN_SET_CARDS, cards });
+export const setFlipped = (array) => ({ type: MAIN_SET_FLIPPED, array });
+export const setDimension = (value) => ({ type: MAIN_SET_DIMENSION, value });
+export const setSolved = (array) => ({ type: MAIN_SET_SOLVED, array });
+export const setDisabled = (boolean) => ({ type: MAIN_SET_DISABLED, boolean });
 
 export const resizeBoard = () => (dispatch) => {
   dispatch(
@@ -76,12 +84,18 @@ export const resizeBoard = () => (dispatch) => {
   );
 };
 
+export const showAllCards = () => (dispatch, getState) => {
+  const cards = getState().main.cards;
+  dispatch(setFlipped(cards.map((card) => card.id)));
+  setTimeout(() => dispatch(setFlipped([])), 3000);
+};
+
 export const handleClick = (id) => (dispatch, getState) => {
-   const cards = getState().main.cards;
-   const flipped = getState().main.flipped;
-   const solved = getState().main.solved;
-   const sameCardClicked = flipped.includes(id);
-   dispatch(setDisabled(true));
+  const cards = getState().main.cards;
+  const flipped = getState().main.flipped;
+  const solved = getState().main.solved;
+  const sameCardClicked = flipped.includes(id);
+  dispatch(setDisabled(true));
 
   if (flipped.length === 0) {
     dispatch(setFlipped([id]));
@@ -91,9 +105,9 @@ export const handleClick = (id) => (dispatch, getState) => {
     dispatch(setFlipped([flipped[0], id]));
     if (isMatch(id, cards, flipped)) {
       dispatch(setSolved([...solved, flipped[0], id]));
-      dispatch(resetCards());
+      resetCards(dispatch);
     } else {
-      setTimeout(dispatch(resetCards), 1000);
+      setTimeout(() => resetCards(dispatch), 1000);
     }
   }
 };
@@ -106,9 +120,18 @@ const isMatch = (id, cards, flipped) => {
   return flippedCard.type === clickedCard.type;
 };
 
-const resetCards = () => (dispatch) => {
-   dispatch(setFlipped([]));
-   dispatch(setDisabled(false));
+const resetCards = (dispatch) => {
+  dispatch(setFlipped([]));
+  dispatch(setDisabled(false));
 };
+
+export const setCountPairs = (count) => (dispatch) => {
+  const cardsArray = initializeDeck(count);
+  dispatch(setCards(cardsArray));
+};
+
+export const succeedGame = (count) => (dispatch) => {
+  dispatch(resetState(count))
+}
 
 export default mainReducer;
